test: adapt secpump fixtures and patch source to LLVM 22 CIR
Merging main brought in four secpump fixtures that pass under LLVM 21
but fail on this branch's merge commit:

- patchir-decomp/secpump_{malicious_memcpy,attribute_modified_cb}: the
  CIR printer in LLVM 22 emits `cir.func dso_local @X(` where it
  previously emitted `cir.func @X(`. Loosen each CIR-LABEL to
  `cir.func {{.*}}@X(` so both forms match (the rest of the
  patchir-decomp suite already uses an explicit `dso_local` literal,
  but `{{.*}}` keeps the fixture forward-compatible without a future
  flag-day update).
- patchir-klee-verifier/secpump_e2e: same shape, in LLVM IR. The
  KLEE-LABELs for `@memcpy_caller` and `@main` now allow `dso_local`
  between `define` and the return type.
- patches/patch_process_vuln_req.c: under LLVM 22 the embedded clang
  frontend skips the usual integer promotions for
  `Attack_It (uint16_t) >= sizeof(AttackBuffer) (size_t)` and emits
  `cir.cmp` with mismatched operand types, failing CIR verification at
  patch-load time and cascading into the patchir-transform and
  patchir-klee-verifier tests. Cast both sides to `unsigned int` so
  the comparison survives the new lowering with identical semantics.

+++ test/patchir-transform/patches/patch_process_vuln_req.c
@@ -35,7 +35,12 @@ void patch__replace__ProcessVulnReq(uint8_t *att_data) {
     static uint16_t Attack_It = 0;
 
     for (int i = 0; i < 16; ++i, ++Attack_It) {
-        if (Attack_It >= sizeof(AttackBuffer)) {
+        // Explicit cast: under LLVM 22 the CIR frontend skips the usual
+        // integer promotion for `uint16_t >= sizeof(...)` and emits a
+        // `cir.cmp` whose operands have differing types, which fails
+        // verification. Casting both sides to a common width keeps the
+        // semantics identical and survives the new lowering.
+        if ((unsigned int)Attack_It >= (unsigned int)sizeof(AttackBuffer)) {
             secpump_assert_fail();
         }
         AttackBuffer[Attack_It] = att_data[i];
